+++ src/routes/(fullscreen)/auth/login/+page.svelte
@@ -5,8 +5,8 @@
   import { TitleDiv, BottomDiv } from "$lib/components/divs";
   import { TextInput } from "$lib/components/inputs";
   import { refreshToken } from "$lib/hooks/callAPI";
-  import { keyPairStore } from "$lib/stores";
-  import { requestLogin } from "./service";
+  import { keyPairsStore } from "$lib/stores";
+  import { requestLogin, requestTokenUpgrade } from "./service";
 
   let { data } = $props();
 
@@ -16,14 +16,23 @@
   const login = async () => {
     // TODO: Validation
 
-    if (await requestLogin(email, password, $keyPairStore)) {
+    try {
+      if (!(await requestLogin(email, password))) throw new Error("Failed to login");
+
+      if (
+        $keyPairsStore &&
+        !(await requestTokenUpgrade($keyPairsStore.encKeyPair, $keyPairsStore.sigKeyPair))
+      )
+        throw new Error("Failed to upgrade token");
+
       await goto(
-        $keyPairStore
+        $keyPairsStore
           ? data.redirectPath
           : "/key/generate?redirect=" + encodeURIComponent(data.redirectPath),
       );
-    } else {
+    } catch (e) {
       // TODO: Alert
+      throw e;
     }
   };
